fix: 修复uid过长的bug

+++ src/BilibiliLiveCommon/Model/LiveRoomInfo.cs
@@ -5,12 +5,12 @@
         /// <summary>
         /// 
         /// </summary>
-        public int room_id { get; set; }
+        public long room_id { get; set; }
 
         /// <summary>
         /// 
         /// </summary>
-        public int uid { get; set; }
+        public long uid { get; set; }
 
         /// <summary>
         ///
